+++ kernel/inc/mm.h
@@ -19,6 +19,8 @@
 #ifndef __MM_H__
 #define __MM_H__
 #include <stdint.h>
+#include <stddef.h>
+
 #include <frame_pool.h>
 #include <page_table.h>
 #include <vm_pool.h>
@@ -32,6 +34,6 @@ void init_kernmem(struct framepool *kfp,
 		struct pagetable *pgt, 
 		struct vmpool *kheap);
 void init_pgt(void) __attribute__((section("PGT_INIT")));
-void *kmalloc(uint32_t size);
+void *kmalloc(size_t size);
 void kfree(uint32_t addr);
 #endif
